# openapi.json
{
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "evaluate_faithfulness": {
            "default": false,
            "description": "Have a judge model score the answer afterwards (costs LLM calls).",
            "title": "Evaluate Faithfulness",
            "type": "boolean"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ChatTurn"
            },
            "maxItems": 20,
            "title": "History",
            "type": "array"
          },
          "query": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 5,
            "description": "How many sources to answer from.",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          },
          "use_reranker": {
            "default": false,
            "description": "Reorder the candidates with the workspace's reranker.",
            "title": "Use Reranker",
            "type": "boolean"
          }
        },
        "required": [
          "query"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "title": "Answer",
            "type": "string"
          },
          "latency_breakdown": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Breakdown"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "number"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "answer",
          "sources",
          "latency_ms"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ChatTurn": {
        "description": "One earlier message. Only user and assistant turns exist: a caller cannot inject a \"system\" turn.",
        "properties": {
          "content": {
            "maxLength": 4000,
            "title": "Content",
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "ChatTurn",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IssueKeyRequest": {
        "properties": {
          "tenant_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9-]{1,64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Tenant to issue a key for. A new UUID is generated when omitted.",
            "title": "Tenant Id"
          }
        },
        "title": "IssueKeyRequest",
        "type": "object"
      },
      "IssueKeyResponse": {
        "properties": {
          "api_key": {
            "title": "Api Key",
            "type": "string"
          },
          "tenant_id": {
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "api_key"
        ],
        "title": "IssueKeyResponse",
        "type": "object"
      },
...
      "RetrievalComparison": {
        "properties": {
          "baseline": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Baseline",
            "type": "array"
          },
          "baseline_latency_ms": {
            "title": "Baseline Latency Ms",
            "type": "number"
          },
          "degraded": {
            "description": "Why the result is not what the configuration asked for.",
            "items": {
              "type": "string"
            },
            "title": "Degraded",
            "type": "array"
          },
          "reranked": {
            "items": {
              "$ref": "#/components/schemas/Source"
            },
            "title": "Reranked",
            "type": "array"
          },
          "reranked_latency_ms": {
            "title": "Reranked Latency Ms",
            "type": "number"
          },
          "reranker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reranker"
          }
        },
        "required": [
          "baseline",
          "reranked",
          "baseline_latency_ms",
          "reranked_latency_ms"
        ],
        "title": "RetrievalComparison",
        "type": "object"
      },
      "RevokeKeysRequest": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Revoke this key.",
            "title": "Api Key"
          },
          "tenant_id": {
            "anyOf": [
              {
                "pattern": "^[A-Za-z0-9-]{1,64}$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Revoke every key of this tenant.",
            "title": "Tenant Id"
          }
        },
        "title": "RevokeKeysRequest",
        "type": "object"
      },
      "RevokeKeysResponse": {
        "properties": {
          "revoked": {
            "title": "Revoked",
            "type": "integer"
          }
        },
        "required": [
          "revoked"
        ],
        "title": "RevokeKeysResponse",
        "type": "object"
      },
      "Source": {
        "properties": {
          "chunk_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chunk Preview"
          },
          "section": {
            "title": "Section",
            "type": "string"
          },
          "similarity_score": {
            "title": "Similarity Score",
            "type": "number"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url",
          "section",
          "similarity_score"
        ],
        "title": "Source",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-Admin-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Nexus: retrieval-augmented question answering over your own documents, with the tools to measure how well it retrieves. Authenticate with `X-API-Key` (a workspace key); `/v1/admin` takes `X-Admin-Key`. Every error has the body `{code, message, request_id}`.",
    "title": "Nexus RAG API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/keys": {
      "post": {
        "description": "Issues a tenant API key. The key is shown once; only its hash is stored.",
        "operationId": "issue_api_key",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IssueKeyRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueKeyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Issue Api Key",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/keys/revoke": {
      "post": {
        "description": "Revokes one key (by its value) or every key of a tenant.",
        "operationId": "revoke_api_keys",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeKeysRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevokeKeysResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Revoke Api Keys",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/chat": {
      "post": {
        "description": "Answers a question from the workspace's documents, with the sources it used.",
        "operationId": "chat",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Chat",
        "tags": [
          "chat"
        ]
      }
    },
    "/v1/chat/stream": {
      "post": {
        "description": "Server-sent events, one JSON object per `data:` line: `token` (answer text, repeated), `sources`, `done` (the answer is complete), then optionally `faithfulness` (seconds later, when requested). A failure while streaming ends the stream with an `error` event; a failure before it starts is an ordinary error response.",
        "operationId": "chat_stream",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Chat Stream",
        "tags": [
          "chat"
        ],
        "x-events": {
          "$defs": {
            "DoneEvent": {
              "properties": {
                "latency_ms": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Latency Ms"
                },
                "type": {
                  "const": "done",
                  "default": "done",
                  "title": "Type",
                  "type": "string"
                }
              },
              "title": "DoneEvent",
              "type": "object"
            },
            "ErrorEvent": {
              "properties": {
                "code": {
                  "title": "Code",
                  "type": "string"
                },
                "message": {
                  "title": "Message",
                  "type": "string"
                },
                "type": {
                  "const": "error",
                  "default": "error",
                  "title": "Type",
                  "type": "string"
                }
              },
              "required": [
                "code",
                "message"
              ],
              "title": "ErrorEvent",
              "type": "object"
            },
            "FaithfulnessContent": {
              "properties": {
                "reasoning": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Reasoning"
                },
                "score": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Score"
                }
              },
              "title": "FaithfulnessContent",
              "type": "object"
            },
            "FaithfulnessEvent": {
              "properties": {
                "content": {
                  "$ref": "#/$defs/FaithfulnessContent"
                },
                "type": {
                  "const": "faithfulness",
                  "default": "faithfulness",
                  "title": "Type",
                  "type": "string"
                }
              },
              "required": [
                "content"
              ],
              "title": "FaithfulnessEvent",
              "type": "object"
            },
            "Source": {
              "properties": {
                "chunk_preview": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Chunk Preview"
                },
                "section": {
                  "title": "Section",
                  "type": "string"
                },
                "similarity_score": {
                  "title": "Similarity Score",
                  "type": "number"
                },
                "url": {
                  "title": "Url",
                  "type": "string"
                }
              },
              "required": [
                "url",
                "section",
                "similarity_score"
              ],
              "title": "Source",
              "type": "object"
            },
            "SourcesEvent": {
              "properties": {
                "content": {
                  "items": {
                    "$ref": "#/$defs/Source"
                  },
                  "title": "Content",
                  "type": "array"
                },
                "type": {
                  "const": "sources",
                  "default": "sources",
                  "title": "Type",
                  "type": "string"
                }
              },
              "required": [
                "content"
              ],
              "title": "SourcesEvent",
              "type": "object"
            },
            "TokenEvent": {
              "properties": {
                "content": {
                  "title": "Content",
                  "type": "string"
                },
                "type": {
                  "const": "token",
                  "default": "token",
                  "title": "Type",
                  "type": "string"
                }
              },
              "required": [
                "content"
              ],
              "title": "TokenEvent",
              "type": "object"
            }
          },
          "description": "Documentation only: the events of POST /v1/chat/stream, one per `data:` line, as JSON.",
          "properties": {
            "done": {
              "$ref": "#/$defs/DoneEvent"
            },
            "error": {
              "$ref": "#/$defs/ErrorEvent"
            },
            "faithfulness": {
              "$ref": "#/$defs/FaithfulnessEvent"
            },
            "sources": {
              "$ref": "#/$defs/SourcesEvent"
            },
            "token": {
              "$ref": "#/$defs/TokenEvent"
            }
          },
          "required": [
            "token",
            "sources",
            "done",
            "faithfulness",
            "error"
          ],
          "title": "ChatStreamEvents",
          "type": "object"
        }
      }
    },
...
    "/v1/retrieval/compare": {
      "post": {
        "description": "What the reranker changes for this question: the same retrieval with and without it. No LLM is called.",
        "operationId": "compare_retrieval",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalComparison"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A valid credential is required."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "A rate limit or quota was reached."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unexpected failure; quote request_id."
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Compare Retrieval",
        "tags": [
          "chat"
        ]
      }
    },
...
  }
}